user class created, checkout data is stored in it
