+++ src/main/webapp/json/data.json
@@ -1,1 +1,1 @@
-{"id":0,"userId":0,"firstName":"Eva","lastName":"Kádár","email":"evi.kadar@gmail.com","address":"Szlatina u. 5.","address2":"3rd floor, appartement 4","state":"Budapest","zip":"1144","country":"Hungary"}
+{"id":0,"userId":0,"firstName":"Eva","lastName":"Kádár","email":"","address":"Szlatina u. 5.","address2":"","state":"","zip":"1144","country":""}
